CMC for search space with only changing crank angle optimal force
See notes in tabs with 3/8/21 naming of:

https://docs.google.com/spreadsheets/d/1E5ZpY3X6rg-SnpGjvgGRdW8wEWR2rKgzLcXOoUtjTak/edit#gid=1579767403

In two of the trials from the #1 tab, the ankle looked snapped back at the start of the motion. To rectify this in the #2 tab, the Setup file had to be altered to start at 0s (instead of 0.3s). However, when trying to use this startup file for the other trials that previously ran, for some reason the solver did not always find a solution. Will review methods tonight at meeting